auth-service/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
sqlalchemy==2.0.25
pydantic[email]==1.10.14
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
